fix nav
