+++ packages/client/src/App.tsx
@@ -1,7 +1,5 @@
 import { useEffect } from 'react'
 import './App.css'
-import { useNavigate } from 'react-router-dom'
-import yApiService from 'services/y-api-service'
 
 function App(): JSX.Element {
   useEffect(() => {
@@ -13,19 +11,6 @@ function App(): JSX.Element {
     }
 
     fetchServerData()
-  }, [])
-
-  const navigate = useNavigate()
-
-  const checkUserAuth = async (): Promise<void> => {
-    const res = await yApiService.getUser()
-    if (!res?.login) {
-      navigate('/sign-in')
-    }
-  }
-
-  useEffect(() => {
-    checkUserAuth()
   }, [])
 
   return <div className="App">Вот тут будет жить ваше приложение :)</div>
