+++ packages/client/src/config/router.tsx
@@ -20,6 +20,19 @@ import { routes } from 'config/routes'
 import yApiService from 'services/y-api-service'
 
 export const AppRouter: React.FC = () => {
+  const navigate = useNavigate()
+
+  const checkUserAuth = async (): Promise<void> => {
+    const res = await yApiService.getUser()
+    if (!res?.login) {
+      navigate('/sign-in')
+    }
+  }
+
+  useEffect(() => {
+    checkUserAuth()
+  }, [])
+
   return (
     <Routes>
       <Route element={<AppLayout />} errorElement={<ServerErrorPage />}>
